+++ package.json
@@ -19,7 +19,7 @@
   "dependencies": {
     "@amplitude/analytics-browser": "^1.2.3",
     "@metamask/post-message-stream": "^5.0.0",
-    "eth-chains": "^0.4.0",
+    "eth-chains": "^1.0.0",
     "eth-rpc-errors": "^4.0.3",
     "ethers": "^5.6.9",
     "extension-port-stream": "^2.0.1",
