+++ src/lib/utils/chains.ts
@@ -1,0 +1,44 @@
+import { ChainId, chains } from 'eth-chains';
+
+// ALL THE BELOW ARE COPIED FROM REVOKE.CASH AND SHOULD BE EXTRACTED AT SOME POINT
+
+export const getChainExplorerUrl = (chainId: number): string | undefined => {
+  const overrides: any = {
+    [ChainId.SmartBitcoinCash]: 'https://smartscan.cash',
+    [ChainId.Moonbeam]: 'https://moonbeam.moonscan.io',
+    [ChainId.Moonriver]: 'https://moonriver.moonscan.io',
+    [ChainId.CeloMainnet]: 'https://celoscan.io',
+    [ChainId.CeloAlfajoresTestnet]: 'https://alfajores.celoscan.io',
+    [ChainId.AuroraMainnet]: 'https://aurorascan.dev',
+    [ChainId.AuroraTestnet]: 'https://testnet.aurorascan.dev',
+    [ChainId.BitTorrentChainMainnet]: 'https://bttcscan.com',
+    [ChainId.BitTorrentChainTestnet]: 'https://testnet.bttcscan.com',
+    [ChainId.CLVParachain]: 'https://clvscan.com',
+    [ChainId.SyscoinTanenbaumTestnet]: 'https://tanenbaum.io',
+    [ChainId.SyscoinMainnet]: 'https://explorer.syscoin.org',
+    [ChainId.Astar]: 'https://blockscout.com/astar',
+    [ChainId.Gnosis]: 'https://gnosisscan.io',
+    [421613]: 'https://goerli.arbiscan.io',
+    [42170]: 'https://nova.arbiscan.io',
+  };
+
+  const [explorer] = chains.get(chainId)?.explorers ?? [];
+
+  return overrides[chainId] ?? explorer?.url;
+};
+
+export const getChainRpcUrl = (chainId: number, infuraKey: string = ''): string | undefined => {
+  // These are not in the eth-chains package, so manually got from chainlist.org
+  const overrides: any = {
+    [ChainId.ArbitrumOne]: 'https://arb1.arbitrum.io/rpc',
+    [421613]: 'https://goerli-rollup.arbitrum.io/rpc',
+    [42170]: 'https://nova.arbitrum.io/rpc',
+    [ChainId.Moonbeam]: 'https://moonbeam.public.blastapi.io',
+    [ChainId.Sepolia]: `https://sepolia.infura.io/v3/${infuraKey}`,
+    [ChainId.Shiden]: 'https://shiden.public.blastapi.io',
+    [ChainId.GodwokenMainnet]: 'https://v1.mainnet.godwoken.io/rpc',
+  };
+
+  const [rpcUrl] = chains.get(chainId)?.rpc ?? [];
+  return overrides[chainId] ?? rpcUrl?.replace('${INFURA_API_KEY}', infuraKey);
+};
